move transform func call test to transform.rs
