+++ ic-http/src/request.rs
@@ -98,47 +98,3 @@ impl Default for HttpRequestBuilder {
         Self::new()
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    /// Transform function which is intentionally creates a new request
-    /// with itself as a transform function.
-    /// This is potentially causing to rewrite the transform function
-    /// in a thread-local storage while it is being executed, which
-    /// leads to a hang.
-    fn transform_function_with_overwrite(arg: TransformArgs) -> HttpResponse {
-        create_request_with_transform();
-        arg.response
-    }
-
-    /// Creates a request with a transform function which overwrites itself.
-    fn create_request_with_transform() -> CanisterHttpRequestArgument {
-        create_request()
-            .url("https://www.example.com")
-            .transform_func(transform_function_with_overwrite, vec![])
-            .build()
-    }
-
-    // IMPORTANT: If this test hangs check the implementation of inserting
-    // transform function to the thread-local storage.
-    //
-    // This test simulates the case when transform function tries to
-    // rewrite itself in a thread-local storage while it is being executed.
-    // This may lead to a hang if the insertion to the thread-local storage
-    // is not written properly.
-    #[tokio::test]
-    async fn test_request_with_nested_transform_insert() {
-        // Arrange
-        let request = create_request_with_transform();
-        let mock_response = crate::response::create_response().build();
-        crate::mock::mock(request.clone(), mock_response);
-
-        // Act
-        let (response,) = crate::mock::http_request(request).await.unwrap();
-
-        // Assert
-        assert_eq!(response.status, 200);
-    }
-}
