+++ ic-http/src/transform.rs
@@ -49,6 +49,7 @@ fn get_function_name<F>(_: &F) -> &'static str {
 #[cfg(test)]
 mod test {
     use super::*;
+    use ic_cdk::api::management_canister::http_request::CanisterHttpRequestArgument;
 
     /// A test transform function.
     fn transform_function_1(arg: TransformArgs) -> HttpResponse {
@@ -82,5 +83,49 @@ mod test {
 
         // Assert.
         assert_eq!(names, vec!["transform_function_1", "transform_function_2"]);
+    }
+
+    /// Transform function which is intentionally creates a new request
+    /// with itself as a transform function.
+    /// This is potentially causing to rewrite the transform function
+    /// in a thread-local storage while it is being executed, which
+    /// leads to a hang.
+    fn transform_function_with_overwrite(arg: TransformArgs) -> HttpResponse {
+        create_request_with_transform();
+        arg.response
+    }
+
+    /// Creates a request with a transform function which overwrites itself.
+    fn create_request_with_transform() -> CanisterHttpRequestArgument {
+        crate::request::create_request()
+            .url("https://www.example.com")
+            .transform_func(transform_function_with_overwrite, vec![])
+            .build()
+    }
+
+    // IMPORTANT: If this test hangs check the implementation of inserting
+    // transform function to the thread-local storage.
+    //
+    // This test simulates the case when transform function tries to
+    // rewrite itself in a thread-local storage while it is being executed.
+    // This may lead to a hang if the insertion to the thread-local storage
+    // is not written properly.
+    #[tokio::test]
+    async fn test_transform_function_call_without_a_hang() {
+        // Arrange
+        let request = create_request_with_transform();
+        let mock_response = crate::response::create_response().build();
+        crate::mock::mock(request.clone(), mock_response);
+
+        // Act
+        let (response,) = crate::mock::http_request(request.clone()).await.unwrap();
+
+        // Assert
+        assert_eq!(response.status, 200);
+        assert_eq!(crate::mock::times_called(request), 1);
+        assert_eq!(
+            crate::mock::registered_transform_function_names(),
+            vec!["transform_function_with_overwrite"]
+        );
     }
 }
